refactor: move models into models module and dto conversion to core
